Fixing double convertions from rational numbers if result value in range (0, 1)

+++ Digimarc.NTiff/Types/Rational.cs
@@ -34,7 +34,7 @@ namespace Digimarc.NTiff.Types
         public readonly uint Denominator;
         public double ToDouble()
         {
-            return Numerator / Denominator;
+            return (double)Numerator / Denominator;
         }
 
         public override string ToString()
